Break out the state machine logic into its own module
