+++ lib/platform-independent/src/state-machine.h
@@ -1,0 +1,31 @@
+#ifndef LIB_STATE_MACHINE_H
+#define LIB_STATE_MACHINE_H
+
+#include <stdint.h>
+
+typedef enum state
+{
+    IDLE,
+    RUNNING,
+    PAUSED,
+    RINGING,
+} state_t;
+
+typedef enum event
+{
+    PRESS,
+    CW_ROTATION,
+    CCW_ROTATION,
+    DOUBLE_PRESS,
+    LONG_PRESS,
+    CW_PRESSED_ROTATION,
+    CCW_PRESSED_ROTATION,
+    TIMEOUT,
+    SECOND_TICK,
+} event_t;
+
+void step_state(event_t event);
+void init_state_machine(void);
+void service_state_machine(void);
+
+#endif // LIB_STATE_MACHINE_H
